Restructure pylot, check if test works on windows
